maint: date bound on old dependencies
The current resolution method for testing with lowest supported
depenencies is lowest-direct. `lowest` is not practical because
various dependencies do not themselves properly check that they work
with the lowest specified versions of their dependencies, with some
packages running down to their first release ever.

So I set --exclude-newer <yesterday's date> such that at least the
indirect dependencies are not updated any more. I tried setting it to
an older date but in the development groups there are recent releases
so that didn't work, so this bound is going to have an effect in the
future.