tests: soft-remove non-load-bearing threshold
The test that checks that jit_and_block_if_profiling does block needs to
check that execution time is at least something. It also included a test
that it was roughly equal to the expected runtime. However it's flaky on
CI, sometimes some overhead appears out of who know where. So now it
just checks the running time isn't insanely large.