perf: share single bart run between slow tests
There are 3 unit tests that require long mcmc chains. Now they share the
same mcmc instead of each running its own. This did not make the tests
actually any faster, but because of an important performance bug that
this optimization effort uncovered. So I commit anyway since it's not
slower, and hopefully will become faster when I fix the performance bug.
It's a serious bug so I predict I will fix it instead of
procrastinating.