perf: benchmark to pin trace duplication bug
This commit adds a benchmark case that measures mcmc loop running time
vs. number of samples to save in the trace. The loop is stopped after a
fixed number of samples, shorter than the planned trace. The mcmc is run
without data. If the running time increases linearly with trace length,
the trace is being duplicated on each iteration. The running time should
be flat if everything is working corretly. It's currently linear.