fix: revert method used to write to trace
A while ago the way mcmc samples are saved to the trace on each
iteration changed from assignment to conditional assignment. This was
intended to simplify the logic of the code and allow arbitrary
operations in the context of saving the trace without worry they would
be duplicated due to saving two traces. This commit switches back to
simple assignment, always assigning to both the burnin and main trace,
because it fixes the bug identified in the last few commits that the
traces would be duplicated on each iteration.