test: change xdist scheduling
The previous scheduling was optimal (on my laptop) with the bugged slow
tests, now it's convenient to switch back to worksteal, which is also
probably going to adapt better in the future to whatever happens.

Now the fixture cachedbart is implemented using pytest fixture scopes
and it's a class fixture, instead of saving to a global dict.