refactor: error_cov_inv_* => error_cov_*
The error_cov_inv_df and error_cov_inv_scale attributes had unintuitive
names because they are the parameters of the distribution of error_cov,
not error_cov_inv. In particular error_cov_inv_scale is a scale of the
*inverse* Wishart distribution on error_cov, not a scale of the Wishart
distribution on the *inverse* error_cov. However on reading the names
the second interpretation is the one that readily comes to mind as
there's an attribute named error_cov_inv.

During the refactor I noticed a related bug: error_cov_inv is
initialized using error_cov_inv_scale as a scale instead of as an
inverse scale.